Added barrelling for main window
